results-domain/results-domain-transformations/results-domain-anonymization-events/src/main/java/uk/gov/moj/cpp/data/anonymization/generator/RegexGenerator.java
package uk.gov.moj.cpp.data.anonymization.generator;

import uk.gov.moj.cpp.data.anonymization.generator.Generator;

import com.mifmif.common.regex.Generex;

import java.util.regex.Pattern;

public class RegexGenerator implements Generator<String> {

    private final Generex generex;

    public RegexGenerator(final Pattern pattern) {
        generex = new Generex(pattern.toString());
    }

    @Override
    public String convert(final String fieldValue) {
        return generex.random();
    }
}